refactor: auto-split test_tracing.py
