+++ tests/unit/api/tracing/test_all_exports.py
@@ -1,0 +1,53 @@
+"""Tests for tracing - Request tracing and correlation ID utilities.
+
+These tests verify the tracing system using Design by Contract
+principles to ensure proper request tracking across the API.
+"""
+
+import re
+from typing import NoReturn
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+# Configure async tests to use asyncio backend only
+pytestmark = pytest.mark.anyio
+
+
+@pytest.fixture(scope="module")
+def anyio_backend() -> str:
+    """Use asyncio backend only (trio not installed)."""
+    return "asyncio"
+
+
+class TestAllExports:
+    """Tests for module __all__ exports."""
+
+    def test_tracing_all_exports_importable(self) -> None:
+        """Test that all __all__ exports are importable."""
+        import src.api.utils.tracing as tracing
+        from src.api.utils.tracing import __all__
+
+        for name in __all__:
+            assert hasattr(tracing, name), f"Missing export: {name}"
+
+    def test_expected_exports_present(self) -> None:
+        """Test that expected exports are in __all__."""
+        from src.api.utils.tracing import __all__
+
+        expected = [
+            "CORRELATION_ID_HEADER",
+            "REQUEST_ID_HEADER",
+            "TraceContext",
+            "RequestTracer",
+            "generate_request_id",
+            "generate_correlation_id",
+            "get_request_id",
+            "set_request_id",
+            "get_trace_context",
+            "set_trace_context",
+            "traced_log",
+        ]
+
+        for name in expected:
+            assert name in __all__, f"Missing from __all__: {name}"
