+++ tests/unit/api/tracing/test_generate_correlation_id_contract.py
@@ -1,0 +1,38 @@
+"""Tests for tracing - Request tracing and correlation ID utilities.
+
+These tests verify the tracing system using Design by Contract
+principles to ensure proper request tracking across the API.
+"""
+
+import re
+from typing import NoReturn
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+# Configure async tests to use asyncio backend only
+pytestmark = pytest.mark.anyio
+
+
+@pytest.fixture(scope="module")
+def anyio_backend() -> str:
+    """Use asyncio backend only (trio not installed)."""
+    return "asyncio"
+
+
+class TestGenerateCorrelationIdContract:
+    """Design by Contract tests for generate_correlation_id function."""
+
+    def test_tracing_returns_string(self) -> None:
+        """Postcondition: Returns a string."""
+        from src.api.utils.tracing import generate_correlation_id
+
+        result = generate_correlation_id()
+        assert isinstance(result, str)
+
+    def test_starts_with_prefix(self) -> None:
+        """Postcondition: Starts with 'cor_' prefix."""
+        from src.api.utils.tracing import generate_correlation_id
+
+        result = generate_correlation_id()
+        assert result.startswith("cor_")
