+++ tests/unit/api/tracing/test_generate_request_id.py
@@ -1,0 +1,39 @@
+"""Tests for tracing - Request tracing and correlation ID utilities.
+
+These tests verify the tracing system using Design by Contract
+principles to ensure proper request tracking across the API.
+"""
+
+import re
+from typing import NoReturn
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+# Configure async tests to use asyncio backend only
+pytestmark = pytest.mark.anyio
+
+
+@pytest.fixture(scope="module")
+def anyio_backend() -> str:
+    """Use asyncio backend only (trio not installed)."""
+    return "asyncio"
+
+
+class TestGenerateRequestId:
+    """Functional tests for generate_request_id."""
+
+    def test_generates_unique_ids(self) -> None:
+        """Test that each call generates a unique ID."""
+        from src.api.utils.tracing import generate_request_id
+
+        ids = {generate_request_id() for _ in range(100)}
+        assert len(ids) == 100  # All unique
+
+    def test_id_format(self) -> None:
+        """Test that ID has expected format."""
+        from src.api.utils.tracing import generate_request_id
+
+        result = generate_request_id()
+        # Format: req_ + 16 hex characters
+        assert re.match(r"^req_[a-f0-9]{16}$", result)
