+++ tests/unit/api/tracing/test_generate_request_id_contract.py
@@ -1,0 +1,45 @@
+"""Tests for tracing - Request tracing and correlation ID utilities.
+
+These tests verify the tracing system using Design by Contract
+principles to ensure proper request tracking across the API.
+"""
+
+import re
+from typing import NoReturn
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+# Configure async tests to use asyncio backend only
+pytestmark = pytest.mark.anyio
+
+
+@pytest.fixture(scope="module")
+def anyio_backend() -> str:
+    """Use asyncio backend only (trio not installed)."""
+    return "asyncio"
+
+
+class TestGenerateRequestIdContract:
+    """Design by Contract tests for generate_request_id function."""
+
+    def test_tracing_returns_string(self) -> None:
+        """Postcondition: Returns a string."""
+        from src.api.utils.tracing import generate_request_id
+
+        result = generate_request_id()
+        assert isinstance(result, str)
+
+    def test_returns_non_empty(self) -> None:
+        """Postcondition: Returns non-empty string."""
+        from src.api.utils.tracing import generate_request_id
+
+        result = generate_request_id()
+        assert len(result) > 0
+
+    def test_starts_with_prefix(self) -> None:
+        """Postcondition: Starts with 'req_' prefix."""
+        from src.api.utils.tracing import generate_request_id
+
+        result = generate_request_id()
+        assert result.startswith("req_")
