+++ tests/unit/api/tracing/test_request_id_context.py
@@ -1,0 +1,75 @@
+"""Tests for tracing - Request tracing and correlation ID utilities.
+
+These tests verify the tracing system using Design by Contract
+principles to ensure proper request tracking across the API.
+"""
+
+import re
+from typing import NoReturn
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+# Configure async tests to use asyncio backend only
+pytestmark = pytest.mark.anyio
+
+
+@pytest.fixture(scope="module")
+def anyio_backend() -> str:
+    """Use asyncio backend only (trio not installed)."""
+    return "asyncio"
+
+
+class TestRequestIdContext:
+    """Functional tests for request ID context management."""
+
+    def test_get_returns_empty_when_not_set(self) -> None:
+        """Test that get returns empty string when not set."""
+        from src.api.utils.tracing import (
+            _request_id_var,
+            get_request_id,
+        )
+
+        # Reset to default state
+        token = _request_id_var.set("")
+        try:
+            result = get_request_id()
+            assert result == ""
+        finally:
+            _request_id_var.reset(token)
+
+    def test_set_and_get_round_trip(self) -> None:
+        """Test that set and get work together."""
+        from src.api.utils.tracing import get_request_id, set_request_id
+
+        test_id = "req_test123"
+        token = set_request_id(test_id)
+        try:
+            assert get_request_id() == test_id
+        finally:
+            from src.api.utils.tracing import _request_id_var
+
+            _request_id_var.reset(token)
+
+    def test_token_can_reset_value(self) -> None:
+        """Test that token can reset to previous value."""
+        from src.api.utils.tracing import (
+            _request_id_var,
+            get_request_id,
+            set_request_id,
+        )
+
+        original = "original_id"
+        new_id = "new_id"
+
+        original_token = set_request_id(original)
+        try:
+            assert get_request_id() == original
+
+            new_token = set_request_id(new_id)
+            assert get_request_id() == new_id
+
+            _request_id_var.reset(new_token)
+            assert get_request_id() == original
+        finally:
+            _request_id_var.reset(original_token)
