+++ tests/unit/api/tracing/test_request_id_context_contract.py
@@ -1,0 +1,38 @@
+"""Tests for tracing - Request tracing and correlation ID utilities.
+
+These tests verify the tracing system using Design by Contract
+principles to ensure proper request tracking across the API.
+"""
+
+import re
+from typing import NoReturn
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+# Configure async tests to use asyncio backend only
+pytestmark = pytest.mark.anyio
+
+
+@pytest.fixture(scope="module")
+def anyio_backend() -> str:
+    """Use asyncio backend only (trio not installed)."""
+    return "asyncio"
+
+
+class TestRequestIdContextContract:
+    """Design by Contract tests for request ID context management."""
+
+    def test_get_returns_string(self) -> None:
+        """Postcondition: get_request_id returns a string."""
+        from src.api.utils.tracing import get_request_id
+
+        result = get_request_id()
+        assert isinstance(result, str)
+
+    def test_set_returns_token(self) -> None:
+        """Postcondition: set_request_id returns a context token."""
+        from src.api.utils.tracing import set_request_id
+
+        token = set_request_id("test_id")
+        assert token is not None
