+++ tests/unit/api/tracing/test_request_tracer_contract.py
@@ -1,0 +1,39 @@
+"""Tests for tracing - Request tracing and correlation ID utilities.
+
+These tests verify the tracing system using Design by Contract
+principles to ensure proper request tracking across the API.
+"""
+
+import re
+from typing import NoReturn
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+# Configure async tests to use asyncio backend only
+pytestmark = pytest.mark.anyio
+
+
+@pytest.fixture(scope="module")
+def anyio_backend() -> str:
+    """Use asyncio backend only (trio not installed)."""
+    return "asyncio"
+
+
+class TestRequestTracerContract:
+    """Design by Contract tests for RequestTracer middleware."""
+
+    def test_tracing_instantiates(self) -> None:
+        """Postcondition: RequestTracer can be instantiated."""
+        from src.api.utils.tracing import RequestTracer
+
+        tracer = RequestTracer()
+        assert tracer is not None
+
+    def test_has_trace_request_method(self) -> None:
+        """Postcondition: RequestTracer has trace_request method."""
+        from src.api.utils.tracing import RequestTracer
+
+        tracer = RequestTracer()
+        assert hasattr(tracer, "trace_request")
+        assert callable(tracer.trace_request)
