+++ tests/unit/api/tracing/test_trace_context.py
@@ -1,0 +1,66 @@
+"""Tests for tracing - Request tracing and correlation ID utilities.
+
+These tests verify the tracing system using Design by Contract
+principles to ensure proper request tracking across the API.
+"""
+
+import re
+from typing import NoReturn
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+# Configure async tests to use asyncio backend only
+pytestmark = pytest.mark.anyio
+
+
+@pytest.fixture(scope="module")
+def anyio_backend() -> str:
+    """Use asyncio backend only (trio not installed)."""
+    return "asyncio"
+
+
+class TestTraceContext:
+    """Functional tests for TraceContext dataclass."""
+
+    def test_tracing_default_values(self) -> None:
+        """Test default values for optional fields."""
+        from src.api.utils.tracing import TraceContext
+
+        context = TraceContext(
+            request_id="req_123",
+            correlation_id="cor_456",
+        )
+
+        assert context.operation == ""
+        assert context.start_time == 0.0
+        assert context.metadata == {}
+
+    def test_custom_operation(self) -> None:
+        """Test setting custom operation."""
+        from src.api.utils.tracing import TraceContext
+
+        context = TraceContext(
+            request_id="req_123",
+            correlation_id="cor_456",
+            operation="GET /api/health",
+        )
+
+        assert context.operation == "GET /api/health"
+
+    def test_to_dict_values(self) -> None:
+        """Test that to_dict returns correct values."""
+        from src.api.utils.tracing import TraceContext
+
+        context = TraceContext(
+            request_id="req_abc",
+            correlation_id="cor_xyz",
+            operation="POST /api/simulate",
+            metadata={"engine": "mujoco"},
+        )
+        result = context.to_dict()
+
+        assert result["request_id"] == "req_abc"
+        assert result["correlation_id"] == "cor_xyz"
+        assert result["operation"] == "POST /api/simulate"
+        assert result["metadata"] == {"engine": "mujoco"}
