+++ tests/unit/api/tracing/test_trace_context_contract.py
@@ -1,0 +1,50 @@
+"""Tests for tracing - Request tracing and correlation ID utilities.
+
+These tests verify the tracing system using Design by Contract
+principles to ensure proper request tracking across the API.
+"""
+
+import re
+from typing import NoReturn
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+# Configure async tests to use asyncio backend only
+pytestmark = pytest.mark.anyio
+
+
+@pytest.fixture(scope="module")
+def anyio_backend() -> str:
+    """Use asyncio backend only (trio not installed)."""
+    return "asyncio"
+
+
+class TestTraceContextContract:
+    """Design by Contract tests for TraceContext dataclass."""
+
+    def test_tracing_to_dict_returns_dict(self) -> None:
+        """Postcondition: to_dict returns a dictionary."""
+        from src.api.utils.tracing import TraceContext
+
+        context = TraceContext(
+            request_id="req_123",
+            correlation_id="cor_456",
+        )
+        result = context.to_dict()
+        assert isinstance(result, dict)
+
+    def test_to_dict_has_required_fields(self) -> None:
+        """Postcondition: to_dict includes required fields."""
+        from src.api.utils.tracing import TraceContext
+
+        context = TraceContext(
+            request_id="req_123",
+            correlation_id="cor_456",
+        )
+        result = context.to_dict()
+
+        assert "request_id" in result
+        assert "correlation_id" in result
+        assert "operation" in result
+        assert "metadata" in result
