+++ tests/unit/api/tracing/test_trace_context_management.py
@@ -1,0 +1,59 @@
+"""Tests for tracing - Request tracing and correlation ID utilities.
+
+These tests verify the tracing system using Design by Contract
+principles to ensure proper request tracking across the API.
+"""
+
+import re
+from typing import NoReturn
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+# Configure async tests to use asyncio backend only
+pytestmark = pytest.mark.anyio
+
+
+@pytest.fixture(scope="module")
+def anyio_backend() -> str:
+    """Use asyncio backend only (trio not installed)."""
+    return "asyncio"
+
+
+class TestTraceContextManagement:
+    """Functional tests for trace context management."""
+
+    def test_get_returns_none_when_not_set(self) -> None:
+        """Test that get returns None when not set."""
+        from src.api.utils.tracing import (
+            _trace_context_var,
+            get_trace_context,
+        )
+
+        # Reset to default state
+        token = _trace_context_var.set(None)
+        try:
+            result = get_trace_context()
+            assert result is None
+        finally:
+            _trace_context_var.reset(token)
+
+    def test_set_and_get_round_trip(self) -> None:
+        """Test that set and get work together."""
+        from src.api.utils.tracing import (
+            TraceContext,
+            _trace_context_var,
+            get_trace_context,
+            set_trace_context,
+        )
+
+        context = TraceContext(
+            request_id="req_test",
+            correlation_id="cor_test",
+        )
+        token = set_trace_context(context)
+        try:
+            result = get_trace_context()
+            assert result == context
+        finally:
+            _trace_context_var.reset(token)
