+++ tests/unit/api/tracing/test_traced_log_contract.py
@@ -1,0 +1,33 @@
+"""Tests for tracing - Request tracing and correlation ID utilities.
+
+These tests verify the tracing system using Design by Contract
+principles to ensure proper request tracking across the API.
+"""
+
+import re
+from typing import NoReturn
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+# Configure async tests to use asyncio backend only
+pytestmark = pytest.mark.anyio
+
+
+@pytest.fixture(scope="module")
+def anyio_backend() -> str:
+    """Use asyncio backend only (trio not installed)."""
+    return "asyncio"
+
+
+class TestTracedLogContract:
+    """Design by Contract tests for traced_log function."""
+
+    def test_does_not_raise(self) -> None:
+        """Postcondition: traced_log does not raise exceptions."""
+        from src.api.utils.tracing import traced_log
+
+        with patch("src.api.utils.tracing.logger"):
+            # Should not raise
+            traced_log("info", "Test message")
+            traced_log("warning", "Warning message", extra_field="value")
